requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pytz==2020.5
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
